MAINT: Remove dav connector and zope dav connector
