+++ core/docs/changelog/davconnector.maint
@@ -1,0 +1,1 @@
+MAINT: Remove obsolete dav connector and zope dav connector
