--- core/src/zeit/connector/dav/__init__.py
@@ -1,1 +1,0 @@
-# Module
