--- core/src/zeit/connector/dav/tests/__init__.py
@@ -1,1 +1,0 @@
-# python package
