+++ core/src/zeit/connector/filesystem.py
@@ -14,12 +14,10 @@ import zope.app.file.image
 import zope.interface
 
 from zeit.cms.content.sources import FEATURE_TOGGLES
-from zeit.connector.connector import CannonicalId
-from zeit.connector.interfaces import ID_NAMESPACE
+from zeit.connector.interfaces import ID_NAMESPACE, CannonicalId
 from zeit.connector.models import ContentWithMetadataColumns as Content
 import zeit.cms.config
 import zeit.cms.content.dav
-import zeit.connector.dav.interfaces
 import zeit.connector.interfaces
 import zeit.connector.resource
