--- core/src/zeit/connector/longrunning.txt
@@ -1,80 +1,0 @@
-Long running transactions
-=========================
-
-Let's test the behaviour of long running transactions. That is we do something
-with the connector, wait quite a while and continue. This used to yield
-BadStatusLine errors.
-
->>> import zeit.connector.interfaces
->>> import zope.component
->>> connector = zope.component.getUtility(zeit.connector.interfaces.IConnector)
-
-
-Enable debugging of the http data:
-
->>> import zeit.connector.dav.davbase
->>> zeit.connector.dav.davbase.DEBUG_CONNECTION = True
-
-Have to force a reconnect, so debug is actually enabled
->>> connector.disconnect()
-
-Remember the connection for later:
-
->>> connection = connector.get_connection()._con
->>> try:
-...     del connector.property_cache['http://xml.zeit.de/testing/']
-... except KeyError:
-...     pass
-
-Get a list:
-
->>> list(connector.listCollection('http://xml.zeit.de/testing/'))
-send: b'HEAD /cms/testing HTTP/1.1\r\nAccept-Encoding: identity\r\nHost: ...\r\nConnection: keep-alive\r\nUser-Agent: zeit.connector/...\r\n\r\n'
-reply: 'HTTP/1.1 301 Moved Permanently\r\n'
-header: Date: ...
-header: Server: ...
-header: Location: http://.../cms/testing/
-header: Keep-Alive: timeout=..., max=...
-header: Connection: Keep-Alive
-header: Content-Type: text/html; charset=iso-8859-1
-send: b'PROPFIND /cms/testing/ HTTP/1.1\r\n...Depth: 1\r\n...User-Agent: zeit.connector/...<?xml version=\'1.0\' encoding=\'UTF-8\'?>\n<ns0:propfind xmlns:ns0="DAV:"><ns0:allprop/></ns0:propfind>'
-reply: 'HTTP/1.1 207 Multi-Status\r\n'
-header: Date: ...
-header: Server: ...
-header: Content-Length: ...
-header: Keep-Alive: timeout=..., max=...
-header: Connection: Keep-Alive
-header: Content-Type: text/xml; charset="utf-8"
-[]
-
-
-So apparently we may only wait 15 seconds before the connection will time out.
-In this case we will silently reconnect. The first request does not reach the
-server(hence the "reply: ''"), then the reconnect happens and the request is
-sent again:
-
->>> import time
->>> time.sleep(16)
->>> connector.invalidate_cache('http://xml.zeit.de/testing/')
-send: b'PROPFIND /cms/testing/ HTTP/1.1...Depth: 1\r\n...User-Agent: zeit.connector/...<?xml version=\'1.0\' encoding=\'UTF-8\'?>\n<ns0:propfind xmlns:ns0="DAV:"><ns0:allprop/></ns0:propfind>'
-reply: ''
-send: b'PROPFIND /cms/testing/ HTTP/1.1\r\n...Depth: 1...User-Agent: zeit.connector/...<?xml version=\'1.0\' encoding=\'UTF-8\'?>\n<ns0:propfind xmlns:ns0="DAV:"><ns0:allprop/></ns0:propfind>'
-reply: 'HTTP/1.1 207 Multi-Status\r\n'
-header: Date: ...
-header: Server: ...
-header: Content-Length: ...
-header: Keep-Alive: timeout=..., max=...
-header: Connection: Keep-Alive
-header: Content-Type: text/xml; charset="utf-8"
->>> list(connector.listCollection('http://xml.zeit.de/testing/'))
-[]
-
-
-We really have reconnected (see also the connect: (...) above):
-
->>> connection == connector.get_connection()._con
-False
-
-Clean up:
-
->>> zeit.connector.dav.davbase.DEBUG_CONNECTION = False
