+++ core/src/zeit/connector/service.zcml
@@ -2,11 +2,6 @@
    xmlns="http://namespaces.zope.org/zope"
    xmlns:zcml="http://namespaces.zope.org/zcml">
 
-  <utility factory=".zopeconnector.factory"
-    zcml:condition="have zeit.connector"/>
-  <subscriber handler=".zopeconnector.invalidate_cache"
-    zcml:condition="have zeit.connector"/>
-
   <utility factory=".connector.tbc_factory"
     zcml:condition="have zeit.connector.nocache"/>
