--- core/src/zeit/connector/stressing.txt
@@ -1,50 +1,0 @@
-Stress tests
-============
-
->>> import zeit.connector.interfaces
->>> import zope.component
->>> connector = zope.component.getUtility(zeit.connector.interfaces.IConnector)
-
-Let's change properties in a folder in parallel. The actual problem arises only
-on parallel delete on which the server just stops putting out XML.
-
->>> base_id = 'http://xml.zeit.de/testing/stresslock'
->>> from io import BytesIO
->>> import datetime
->>> import pytz
->>> import zeit.connector.interfaces
->>> from zeit.connector.resource import Resource
->>> thread_counter = 0
->>> import threading
->>> lock = threading.Lock()
->>> def stress():
-...     global thread_counter
-...     lock.acquire()
-...     thread_counter += 1
-...     nr = thread_counter
-...     lock.release()
-...     id = base_id + str(nr)
-...     res = Resource(id,
-...                'stresslock',
-...                'text',
-...                BytesIO(b'Pop goes the weasel!'))
-...     res.properties[('foo', 'bar')] = 'baz'
-...     res.properties[('count', 'bar')] = '-1'
-...     until = (
-...         datetime.datetime.now(pytz.UTC) + datetime.timedelta(seconds=120))
-...     token = connector.lock(res.id, 'zope.user', until)
-...     connector.add(res)
-...     for count in range(50):
-...         r = connector[id]
-...         assert r.properties[('count', 'bar')]  == str(count-1), dict(r.properties)
-...         connector.changeProperties(id, {('count', 'bar'): str(count)})
-...     del connector[id]
-...     connector.disconnect()
-
->>> threads = []
->>> while len(threads) < 10:
-...     threads.append(threading.Thread(target=stress))
->>> for t in threads:
-...     t.start()
->>> for t in threads:
-...     t.join()
