+++ core/src/zeit/connector/tests/test_doctest.py
@@ -4,8 +4,7 @@ import zeit.connector.testing
 def test_suite():
     return zeit.connector.testing.FunctionalDocFileSuite(
         'cache.txt',
-        'functional.txt',
-        'invalidation-events.txt',
         'search.txt',
-        layer=zeit.connector.testing.ZOPE_DAV_CONNECTOR_LAYER,
+        'uuid.txt',
+        layer=zeit.connector.testing.MOCK_CONNECTOR_LAYER,
     )
