--- core/src/zeit/connector/tests/test_doctest3.py
@@ -1,8 +1,0 @@
-import zeit.connector.testing
-
-
-def test_suite():
-    return zeit.connector.testing.FunctionalDocFileSuite(
-        'uuid.txt',
-        layer=zeit.connector.testing.MOCK_CONNECTOR_LAYER,
-    )
